requirements.txt
openai==1.30.5
python-dotenv==1.0.1
qdrant-client==1.9.1
sentence-transformers==3.0.0
streamlit==1.35.0
# streamlit-audiorec==0.1.3
transformers==4.41.2
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain_openai
langchain-qdrant